refactor(backend): split hayro backend module
